+++ lib/cocoapods-binary-cache/hooks/post_install.rb
@@ -14,7 +14,7 @@ module PodPrebuild
     def diagnose
       Pod::UI.section("Diagnosing cocoapods-binary-cache") do
         PodPrebuild::Diagnosis.new(
-          cache_validation: PodPrebuild::StateStore.cache_validation,
+          cache_validation: PodPrebuild.state.cache_validation,
           standard_sandbox: @installer_context.sandbox,
           specs: @installer_context.umbrella_targets.map(&:specs).flatten
         ).run