+++ lib/cocoapods-binary-cache/state_store.rb
@@ -1,11 +1,21 @@
 module PodPrebuild
-  class StateStore
-    @excluded_pods = Set.new
-    @cache_validation = CacheValidationResult.new
+  def self.state
+    @state ||= State.new
+  end
 
-    class << self
-      attr_accessor :excluded_pods
-      attr_accessor :cache_validation
+  class State
+    def initialize
+      @store = {
+        :cache_validation => CacheValidationResult.new
+      }
+    end
+
+    def update(data)
+      @store.merge!(data)
+    end
+
+    def cache_validation
+      @store[:cache_validation]
     end
   end
 end